python/pyproject.toml
[tool.poetry]
name = "sidekwest"
version = "0.1.0"
description = ""
authors = ["kwest <a.simple.kwest@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
py-cord = {extras = ["speed"], version = "^2.4.1"}
click = "^8.1.3"
aiofiles = "^23.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^23.3.0"
mypy = "^1.3.0"
pylint = "^2.17.4"

[tool.poetry.scripts]
sidekwest = "sidekwest.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
